requirements.txt
# tensorflow>=2.11.*
tensorflow_datasets
pysam
pandas
click
tqdm
pyyaml